Test and fix statement terminator handling
For some reason I had thought that Emily was Python-like, that newlines
weren't statement terminators inside paren groups, which is dumb
because, then what would semicolons mean inside paren groups?

Now paren groups are parsed as a list of statements. Similar to
Closure's without a .body, they start out in the transient state of
being an 'OpenGroup'; if they aren't a 'Group' by codegen time, an
error is thrown.

This means that now, statement terminators after an expression always
cause the expression to be treated as a statement, either appended to
the statement list of the current group, if any, elsewise emitted as a
statement.

Meanwhile in codegen land, we gotta deal with expressions that may
actually be a sequence of expressions, all but the last of whose values
we want to ignore. Lucky us JavaScript has an obscure operator, the
[comma operator][], that does just that. That's right, the comma is an
operator, and if you're wondering how easily that gets confused with
passing multiple arguments to a function call (in which case you
definitely *don't* want to ignore any of those values), the answer is,
very easily, and in fact we have to add an extra layer of parentheses to
the codegen to disambiguate exactly that.

[comma operator]: http://en.wikipedia.org/wiki/Comma_operator