Quick-and-dirty test script in Bash
Sorry about it being in Bash, but look, it's really not that bad.
We're testing a command-line tool, a command-line shell script makes
plenty of sense. Literally just loop over the tests and diff against
their expected outputs.

Also sorry it's not plain Bourne or even POSIX shell, [process
substitution][] is just such a convenient Bash-ism.

[process substitution]: http://en.wikipedia.org/wiki/Process_substitution