Use built-in Emily value instead of undefined for null
Previously, the JavaScript primitive value `undefined` was used for the
Emily null value. Now, a new built-in value is defined by the runtime
and assigned to the global variable 'null'; like any global variable,
that name can be overridden and everything needs to keep working.

Convenient for ensuring that builtins keep working even if 'null' is
overridden, there's a syntactic construct that evaluates to the null
value [1] that can't be overridden, assuming sane scoping in mcclure's
forthcoming Emily macro system.

[1]: https://bitbucket.org/runhello/emily/src/stable/doc/manual.md#markdown-header-_3

The new built-in null value now stringifies consistent with the Emily
reference implementation ("<null>"), plus it will be able to respond to
comparison methods like .eq and .has etc. I think that makes it worth
the extra hassle of built-in functions having to explicitly return it
(instead of relying on the JS implicit return value of `undefined`).